fix: give example tools a/b separate routes dirs to avoid pod route collision
Pod routes are keyed by file path, so two pods sharing one routesDir emit
duplicate keys and the later pod overwrites the earlier one — tool "a" 404'd
while "b" worked. Move the shared page components into shared.tsx and give
each tool its own routes/ dir (pod/example-tool/<id>/routes) whose files
re-export the shared components.
