+++ docs/plugins/example-tools-pod.ts
@@ -11,6 +11,11 @@ const POD_DIR = path.resolve(__dirname, '../pod/example-tool');
  * query params, and hashes) so navigating around should be mirrored into the
  * CMS address bar as `/cms/tools/<id>/...`. Register two of these (e.g. `a` and
  * `b`) to also exercise switching between tools.
+ *
+ * Each id has its own `routes/` dir (`pod/example-tool/<id>/routes`) because
+ * pod routes are keyed by file path, so two pods sharing one routes dir would
+ * collide and one would 404. The route files just re-export the shared page
+ * components.
  */
 export function exampleToolPod(id: string): Plugin {
   return {
@@ -18,7 +23,7 @@ export function exampleToolPod(id: string): Plugin {
     pod: {
       name: `example-tool-pod-${id}`,
       mount: `/cms-tools/${id}`,
-      routesDir: path.join(POD_DIR, 'routes'),
+      routesDir: path.join(POD_DIR, id, 'routes'),
     },
   };
 }
