+++ docs/pod/example-tool/a/routes/index.tsx
@@ -1,0 +1,1 @@
+export {ExampleToolHome as default} from '../../shared.js';
