+++ docs/pod/example-tool/a/routes/section-one/details/index.tsx
@@ -1,0 +1,1 @@
+export {ExampleToolSectionOneDetails as default} from '../../../../shared.js';
