+++ docs/pod/example-tool/a/routes/section-one/index.tsx
@@ -1,0 +1,1 @@
+export {ExampleToolSectionOne as default} from '../../../shared.js';
