+++ docs/pod/example-tool/b/routes/index.tsx
@@ -1,0 +1,1 @@
+export {ExampleToolHome as default} from '../../shared.js';
