+++ docs/pod/example-tool/b/routes/section-one/index.tsx
@@ -1,0 +1,1 @@
+export {ExampleToolSectionOne as default} from '../../../shared.js';
